updating the topo generator to include cache capacities in the configuration

+++ src/ccn/entity/Router.java
@@ -18,7 +18,7 @@ public class Router extends Node {
 	protected NetworkStack stack;
 	private static final Logger logger = Logger.getConsoleLogger(Router.class.getName());
 
-	public Router(String identity, Point location, List<String> interfaces, int capacity) {
+	public Router(String identity, Point location, List<String> interfaces, long capacity) {
 		super(identity, location, interfaces);
 		stack = RouterStackFactory.buildDefault(this, capacity);
 	}
