+++ src/ccn/entity/stack/RouterStackFactory.java
@@ -4,7 +4,7 @@ import ccn.entity.Router;
 
 public class RouterStackFactory {
 
-	public static NetworkStack buildDefault(Router router, int capacity) {
+	public static NetworkStack buildDefault(Router router, long capacity) {
 		return NetworkStack.buildDefaultNetworkStackWithFiniteCacheCapacity(router, capacity);
 	}
 }
